feat: add mobile responsive design with Carbon UIShell pattern
Adopt Carbon's built-in responsive HeaderContainer + SideNav pattern
to make the app usable on mobile viewports. The sidebar now auto-collapses
below the lg breakpoint (1056px) with a hamburger menu toggle, overlay
sidebar, and backdrop-close behavior — all from Carbon's native CSS.

- Merge AppHeader + AppSidebar into unified responsive shell component
- Add hamburger menu with auto-close on route change
- Align PageShell breakpoints to Carbon lg (66rem)
- Add mobile E2E test tier (TestE2eMobile) with Navigation + Screenshots
- Add i18n keys for menu toggle labels (en + ja)
