+++ .nuke/build.schema.json
@@ -67,6 +67,7 @@
         "TestE2eArticles",
         "TestE2eAuth",
         "TestE2eFeed",
+        "TestE2eMobile",
         "TestE2eMultitenancy",
         "TestServer",
         "TestServerPostmanArticle",
