+++ App/Client/src/App.tsx
@@ -4,7 +4,6 @@ import { FeatureFlagProvider } from './context/FeatureFlagContext';
 import { ProtectedRoute } from './components/ProtectedRoute';
 import { RoleProtectedRoute } from './components/RoleProtectedRoute';
 import { AppHeader } from './components/AppHeader';
-import { AppSidebar } from './components/AppSidebar';
 import { DashboardPage } from './pages/DashboardPage';
 import { LoginPage } from './pages/LoginPage';
 import { RegisterPage } from './pages/RegisterPage';
@@ -19,7 +18,6 @@ function App() {
       <AuthProvider>
         <FeatureFlagProvider>
         <AppHeader />
-        <AppSidebar />
         <Routes>
           <Route
             path="/"
