+++ App/Client/src/i18n/locales/en.json
@@ -7,7 +7,9 @@
     "signIn": "Sign in",
     "signUp": "Sign up",
     "logOut": "Log out",
-    "sideNavLabel": "Side navigation"
+    "sideNavLabel": "Side navigation",
+    "openMenu": "Open menu",
+    "closeMenu": "Close menu"
   },
   "login": {
     "title": "Sign in",
