+++ App/Client/src/index.css
@@ -23,7 +23,13 @@ body {
 #root {
   min-height: 100vh;
   padding-top: var(--header-height);
-  padding-left: var(--sidebar-width);
+}
+
+/* Carbon's SideNav with isChildOfHeader auto-pushes content at lg+ */
+@media (min-width: 66rem) {
+  #root {
+    padding-left: var(--sidebar-width);
+  }
 }
 
 a {
@@ -58,4 +64,3 @@ h1, h2, h3, h4, h5, h6 {
   max-width: 100%;
   padding: 0 15px;
 }
-
