+++ Test/e2e/docker-compose.Mobile.yml
@@ -1,0 +1,3 @@
+services:
+  playwright:
+    command: ["-trx", "/Reports/Test/e2e/Results/e2e-results.trx", "-namespace", "E2eTests.Tests.Mobile"]
